+++ eccorce_app/app/src/main/java/com/example/eccomerce_app/model/ProductModel.kt
@@ -6,13 +6,13 @@ import java.util.UUID
 data class ProductModel(
     val id: UUID,
     val name: String,
-    val description:String,
-    val thumbnail:String,
+    val description: String,
+    val thumbnail: String,
     val subcategoryId: UUID,
     val storeId: UUID,
     val categoryId: UUID,
-    val price: Double,
+    val price: Int,
     val symbol: String,
-    val productVariants:List<List<ProductVariant>>?=null,
-    val productImages:List<String>
+    val productVariants: List<List<ProductVariant>>? = null,
+    val productImages: List<String>
 )
